requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
zope.interface==5.4.0
